MOIDA-54 AWS RDS 생성 뒤 database파일과 연동 및 MakeGroup 일정부분 구조 완성
