+++ database/database.json
@@ -1,0 +1,7 @@
+{
+    "host": "moida-instance.csnx9owbfgoh.ap-northeast-2.rds.amazonaws.com",
+    "user": "admin",
+    "password": "mzmoida5",
+    "port": "3306",
+    "database": "MOIDA_database"
+}
