+++ server/server.js
@@ -1,3 +1,4 @@
+const fs = require('fs');
 const express = require('express');
 const app = express();
 
@@ -5,3 +6,17 @@ const groupMember = require('./Routes/GroupMember');
 app.use('/groupMember', groupMember);
 
 app.listen(5000, () => console.log('express server is listening on port 5000'));
+
+const data=fs.readFileSync('/.database.json');
+const conf = JSON.parse(data);
+const mysql = require('mysql');
+
+const connection = mysql.createConnection({
+    host: conf.host,
+    user: conf.user,
+    password: conf.password,
+    port: conf.port,
+    database: conf.database
+})
+connection.connect();
+
