+++ src/index.jsx
@@ -5,8 +5,7 @@ import { Route, Switch } from 'react-router';
 import { BrowserRouter as Router } from 'react-router-dom';
 import LandingPage from './Page/LandingPage/LandingPage';
 import SignIn from './Page/SignIn/SignIn';
-import PopupDom from './Page/MakeGroup/PopupDom';
-import Main from './Components/Popup/Popup';
+import Main from './Page/Main/Main';
 import MakeGroupPopupContent from './Page/MakeGroup/MakeGroupPopup';
 
 
@@ -18,10 +17,9 @@ const Root = () => {
           <Route exact path="/" component={LandingPage}></Route>
           <Route exact path="/SignIn" component={SignIn}></Route>
           <Route exact path="/SignIn" component={SignIn}></Route>
-          <Route exact path="/Main" component={Main}></Route>
-          
+          <Route exact path="/Main" component={Main}></Route>         
           <Route exact path="/MakeGroupPopupContent" component={MakeGroupPopupContent}></Route>
-
+        
         </Switch>
       </div>
     </Router>
